Refactor user context and form handling logic
Moved user context logic to a new UserContext.jsx and implemented a UserProvider for global state. Wrapped App in UserProvider in main.jsx to provide userContext to App.jsx. Added a custom useFetch1 hook for API requests (only in authentication). Refactored components to use the new context and form logic.